chore: moved settings files
