+++ src/components/features/settings/account-management/components/sendEmail.tsx
