+++ src/components/organisms/tasks/github-connection-card.tsx
@@ -1,6 +1,6 @@
 import useTranslation from 'next-translate/useTranslation';
 
-import { ConnectionHandlerGithub } from '@/components/templates/settings/connected-accounts/connectors/github-connection';
+import { ConnectionHandlerGithub } from '@/components/features/settings/connected-accounts/connectors/github-connection';
 
 import { Button, Stack, Typography } from '@mui/material';
