+++ src/pages/settings/account-management.tsx
@@ -1,7 +1,7 @@
+import { Settings } from '@/components/features/settings';
+import { AccountManagementSettings } from '@/components/features/settings';
 import { HeadContainer } from '@/components/molecules/head-container';
 import { DashboardTemplate } from '@/components/templates/dashboard';
-import { SettingsTemplate } from '@/components/templates/settings';
-import { AccountManagementSettings } from '@/components/templates/settings';
 import { useAuth } from '@/providers/auth';
 
 export default function AccountManagementSettingsPage() {
@@ -18,9 +18,9 @@ export default function AccountManagementSettingsPage() {
           height: '100%',
         }}
       >
-        <SettingsTemplate>
+        <Settings>
           <AccountManagementSettings />
-        </SettingsTemplate>
+        </Settings>
       </DashboardTemplate>
     </>
   ) : null;
