+++ src/pages/settings/developer-portal.tsx
@@ -1,18 +1,11 @@
-import { useRouter } from 'next/router';
-
+import { Settings } from '@/components/features/settings';
+import { DeveloperSettings } from '@/components/features/settings/developer';
 import { HeadContainer } from '@/components/molecules/head-container';
 import { DashboardTemplate } from '@/components/templates/dashboard';
-import { SettingsTemplate } from '@/components/templates/settings';
-import { DeveloperSettings } from '@/components/templates/settings/developer';
-import { ROUTES } from '@/constants/routes';
-import { useUploadImage } from '@/hooks/use-upload-image';
 import { useAuth } from '@/providers/auth';
-import { useMutation } from '@tanstack/react-query';
 
 export default function DeveloperSettingsPage() {
-  const uploadImage = useUploadImage();
-  const { me, gqlAuthMethods, onInvalidateMe } = useAuth();
-  const router = useRouter();
+  const { me } = useAuth();
 
   return me?.id ? (
     <>
@@ -25,9 +18,9 @@ export default function DeveloperSettingsPage() {
           height: '100%',
         }}
       >
-        <SettingsTemplate>
+        <Settings>
           <DeveloperSettings />
-        </SettingsTemplate>
+        </Settings>
       </DashboardTemplate>
     </>
   ) : null;
